requirements.txt
requests==2.31.0
selenium==4.20.0
python-dotenv==1.0.1